kinda fixed the sp02 reading although still unstalbe

+++ IoTM_/STM32CubeIDE/App/sPulsOx/max3010x.cpp
@@ -408,7 +408,7 @@ uint8_t MAX3010x::getRevisionID() {
 // Sample rate = 50
 //Use the default setup if you are just getting started with the MAX30105 sensor
 void MAX3010x::setup(uint8_t powerLevel, uint8_t sampleAverage, uint8_t ledMode,
-		uint16_t sampleRate, uint16_t pulseWidth, uint16_t adcRange) {
+		int sampleRate, uint16_t pulseWidth, uint16_t adcRange) {
 	softReset(); //Reset all configuration, threshold, and data registers to POR values
 
 	//FIFO Configuration
